Make suggested changes to the api
Have altered the controller index endpoint ot be more efficient and
changed the migration for book description to be of type text. As there
is no prod database it is simple to change the migration and
drop/migrate local db to see changes. For a prod database it would
require a new migration.